Rename branch() to this_branch() then alias branch
Needed an alias for git branch akin to those fo git status, git commit,
git add. The sequence would lead to alias="git branch" but the function
branch() already is in use.

Rename the current function to this_branch() and update the PS1
assignments.

Modify get-logs to stop wx-daemon if it's running so that the logs to
capture are in a quiescent state.